animal.all

+++ src/main/java/models/Item.java
@@ -17,7 +17,7 @@ public class Item {
         this.energy = itemType.getEnergy();
     }
 
-    public Item(ItemType itemType,float factor) {
+    public Item(ItemType itemType,double factor) {
         this.itemType = itemType;
         this.price = (int) (itemType.getPrice()*factor);
         this.isEdible = itemType.isEdible();
